% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigtcr_func.R
\name{get.kendalltau}
\alias{get.kendalltau}
\title{Kendall's tau Estimation}
\usage{
get.kendalltau(v, w)
}
\arguments{
\item{v}{Vector of numeric values. Missing values will be ignored.}

\item{w}{vector of numeric values. Missing values will be ignored.}
}
\description{
Estimate Kendall's tau association between two random variables
}
\examples{

kt <- get.kendalltau(pancancer$v, pancancer$w);

}

